* LoD deviations ready

+++ provenance/include/lamure/pro/data/entities/MetaData.h
@@ -11,7 +11,7 @@ class MetaData
     const vec<uint8_t> &get_metadata() const { return _metadata; }
     MetaData() {}
 
-    virtual void read_metadata(ifstream &is, uint32_t meta_data_length)
+    virtual void read_metadata(ifstream &is, uint64_t meta_data_length)
     {
         vec<uint8_t> byte_buffer(meta_data_length, 0);
         is.read(reinterpret_cast<char *>(&byte_buffer[0]), meta_data_length);
